python/requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
sqlalchemy
pyodbc
fastapi
uvicorn
pydantic
python-dotenv
shap
